+++ CHANGELOG.md
@@ -5,6 +5,20 @@ All notable changes to the "rascript-language-server" extension will be document
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [[0.X.Y](https://github.com/joshraphael/rascript-language-server/releases/tag/v0.X.Y)] - 2026-09-DD
+
+[diff](https://github.com/joshraphael/rascript-language-server/compare/v0.5.0...v0.X.Y)
+
+### Added
+
+- Add RATools Error Diagnostics
+
+### Changed
+
+### Removed
+
+### Fixed
+
 ## [[0.5.0](https://github.com/joshraphael/rascript-language-server/releases/tag/v0.5.0)] - 2026-09-18
 
 [diff](https://github.com/joshraphael/rascript-language-server/compare/v0.4.1...v0.5.0)